+++ public/templates/profile/profile.css
@@ -87,6 +87,7 @@ input[type="file"] {
     border-radius: 50%;
     filter: drop-shadow(0px 0px 10px rgba(0, 0, 0, 0.25));
     box-sizing: content-box;
+    flex-shrink: 0;
 }
 
 .change-avatar-button {
